// types.ts
export enum ToneOfVoice {
  Professional = "Professional",
  Friendly = "Friendly",
  Urgent = "Urgent",
  Informative = "Informative",
  Playful = "Playful",
}

export enum ActionType {
  Subject = "Subject",
  Body = "Body",
  AbTest = "AbTest",
  Personalization = "Personalization",
  Cta = "Cta",
  Spam = "Spam",
}